Updated fault model testing for NumberOfCasts

+++ PluginProject/BlackBoxTestCases/NumberOfCastsCheck/NumberOfCastsCheck3.java
@@ -1,4 +1,4 @@
-// This test case covers when there are external method calls and internal method calls.
+// This test is used to find faults counting the number of casts in code, and in comments
 // Total of: 3
 
 class NumberOfCastsCheck3 { // total of: 3
@@ -7,7 +7,7 @@ class NumberOfCastsCheck3 { // total of: 3
         double d = 1.2, e = 3.5, f = 34.1;
         a = (int) d;
         e = (double) b;
-        c = a;
+        c = (double) a;
         // a = (int) d;
         // e = (double) b;
         // c = a;
